# backend-dotnet/api/openapi.yaml
openapi: 3.1.0
info:
  title: Claims API (life claim slice)
  version: 0.2.0
  summary: Individual life, disability income and annuity claims. This version covers one life claim from intake to closed (decision, payment run, close).
  description: |
    Derived from the React mock's `web/src/api` modules (their doc comments carry the verbs). Where the mock's shape is
    unsuitable for a real API the difference is deliberate; the full list is in `backend/README.md` ("Deviations from the mock").

    **Conventions**

    * **Ids** are UUIDs. Claims also have a human `claimNumber` (`L-26-043310`), unique, filterable, never used in paths.
    * **Money** is `{ "amount": "200000.00", "currency": "USD" }`: a decimal string with at most 2 decimals, never a JSON number.
    * **Timestamps** are RFC 3339 in UTC (`2026-09-25T15:03:00Z`). Business dates (date of death, issue date) are plain dates.
    * **Enum values** are snake_case, the same strings the database stores (and the deadline kinds in the architecture doc).
    * **Idempotency**: POSTs that create something take an `Idempotency-Key` header (8 to 128 characters). A repeat with the
      same key and body returns the resource it created (200, `Idempotent-Replayed: true`); the same key with a different body is 409.
    * **Optimistic concurrency**: every mutable resource has a `version`, returned as its `ETag` (`"3"`). State-changing calls on
      an existing resource require `If-Match`: missing is 428, stale is 412.
...
servers:
...
security:
  - bearerAuth: []
paths:
  /claims/life-intake:
    post:
      tags: [Claims]
      operationId: createLifeClaim
      summary: Create a claim from a life intake (notice of death)
      description: |
        One transaction saves the claim, parties, policy snapshot, benefit lines, requirements, the first deadline rows, history
        and an outbox event, then returns `201` with the claim in status `received`. No workflow has run yet: the outbox relay
        starts `orch-<claimNumber>-intake` afterwards. (Mock: `submitLifeIntake` does all of it inline.)
      x-slice-status: implemented
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
        - $ref: '#/components/parameters/Actor'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LifeIntakeRequest'
      responses:
        '201':
          description: Created. `Location` is the claim.
          headers:
            Location: { schema: { type: string } }
            ETag: { $ref: '#/components/headers/ETag' }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Claim' }
        '200':
          description: The same Idempotency-Key and body were already accepted; the claim it created.
          headers:
            Idempotent-Replayed: { schema: { type: string, const: 'true' } }
            ETag: { $ref: '#/components/headers/ETag' }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Claim' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '409': { $ref: '#/components/responses/Conflict' }
        '422': { $ref: '#/components/responses/Unprocessable' }
  /claims:
    get:
      tags: [Claims]
      operationId: listClaims
      summary: List claims, newest first
      description: Mock `listClaims(ownerId)` and `searchClaims(q)` (free-text search is not built).
      x-slice-status: implemented
      parameters:
        - { name: owner, in: query, schema: { type: string, format: uuid } }
        - { name: status, in: query, schema: { $ref: '#/components/schemas/ClaimStatus' } }
        - { name: family, in: query, schema: { $ref: '#/components/schemas/Family' } }
        - { name: claimNumber, in: query, schema: { type: string, pattern: '^[LDA]-[0-9]{2}-[0-9]{6}$' } }
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Cursor'
      responses:
        '200':
          description: A page of claim summaries.
          content:
            application/json:
              schema:
                type: object
                required: [items, nextCursor]
                properties:
                  items: { type: array, items: { $ref: '#/components/schemas/ClaimSummary' } }
                  nextCursor: { type: [string, 'null'] }
        '400': { $ref: '#/components/responses/BadRequest' }
  /claims/{claimId}:
    parameters: [{ $ref: '#/components/parameters/ClaimId' }]
    get:
      tags: [Claims]
      operationId: getClaim
      summary: Get a claim
      description: |
        The resource is the claim's facts and state. The mock's presentation fields (`tags`, `facts`, `stage`, `clocks`, `nextStep`,
        `summary`, `sections`, `exceptions`, `suggestions`) are projections, not stored state, and are not part of this resource;
        `stage` derives from `status`, `clocks` from `deadlines`.
      x-slice-status: implemented
      responses:
        '200':
          description: The claim.
          headers: { ETag: { $ref: '#/components/headers/ETag' } }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Claim' }
        '404': { $ref: '#/components/responses/NotFound' }
  /claims/{claimId}/requirements:
    parameters: [{ $ref: '#/components/parameters/ClaimId' }]
    get:
      tags: [Requirements]
      operationId: listClaimRequirements
      summary: Requirements on a claim
      description: Mock `listRequirements(claimId)`.
      x-slice-status: implemented
      responses:
        '200':
          description: All requirements, oldest first.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/RequirementList' }
        '404': { $ref: '#/components/responses/NotFound' }
...
  /claims/{claimId}/deadlines:
    parameters: [{ $ref: '#/components/parameters/ClaimId' }]
    get:
      tags: [Deadlines]
      operationId: listClaimDeadlines
      summary: Deadline rows for a claim
      description: Mock `LifeFlow.deadlines` in `GET /claims/{id}/workflow`, split into its own resource.
      x-slice-status: implemented
      responses:
        '200':
          description: All rows, oldest first.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/DeadlineList' }
        '404': { $ref: '#/components/responses/NotFound' }
  /deadlines:
    get:
      tags: [Deadlines]
      operationId: listDeadlines
      summary: Deadlines across claims (operations view)
      description: What is open, overdue, or dispatched and stuck. Ordered by `dueAt`.
      x-slice-status: implemented
      parameters:
        - { name: state, in: query, schema: { $ref: '#/components/schemas/DeadlineState' } }
        - { name: kind, in: query, schema: { $ref: '#/components/schemas/DeadlineKind' } }
        - { name: dueBefore, in: query, schema: { type: string, format: date-time } }
        - { name: limit, in: query, schema: { type: integer, minimum: 1, maximum: 500, default: 100 } }
      responses:
        '200':
          description: Deadline rows.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/DeadlineList' }
        '400': { $ref: '#/components/responses/BadRequest' }
  /deadlines/{deadlineId}:
    parameters: [{ $ref: '#/components/parameters/DeadlineId' }]
    get:
      tags: [Deadlines]
      operationId: getDeadline
      summary: Get a deadline row
      description: One deadline row, with its ETag for a later extension.
      x-slice-status: implemented
      responses:
        '200':
          description: The row.
          headers: { ETag: { $ref: '#/components/headers/ETag' } }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Deadline' }
        '404': { $ref: '#/components/responses/NotFound' }
  /deadlines/{deadlineId}:extend:
    parameters: [{ $ref: '#/components/parameters/DeadlineId' }]
    post:
      tags: [Deadlines]
      operationId: extendDeadline
      summary: Extend an open deadline, with a reason
      description: A row update, not a new timer. `dueAt` moves, `originalDueAt` stays, the history records why.
      x-slice-status: implemented
      parameters:
        - $ref: '#/components/parameters/IfMatch'
        - $ref: '#/components/parameters/Actor'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [newDueAt, reason]
              properties:
                newDueAt: { type: string, format: date-time }
                reason: { type: string, minLength: 1 }
      responses:
        '200':
          description: The extended row.
          headers: { ETag: { $ref: '#/components/headers/ETag' } }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Deadline' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
        '412': { $ref: '#/components/responses/PreconditionFailed' }
        '422': { $ref: '#/components/responses/Unprocessable' }
        '428': { $ref: '#/components/responses/PreconditionRequired' }
...
  /claims/{claimId}/letters:
    parameters: [{ $ref: '#/components/parameters/ClaimId' }]
    get:
      tags: [Letters]
      operationId: listClaimLetters
      summary: Outbound letters and messages
      description: Mock `GET /claims/{id}/communications`, outbound only. Calls and inbound items are history events and documents.
      x-slice-status: implemented
      responses:
        '200':
          description: Newest first.
          content:
            application/json:
              schema:
                type: object
                required: [items, nextCursor]
                properties:
                  items: { type: array, items: { $ref: '#/components/schemas/Letter' } }
                  nextCursor: { type: [string, 'null'] }
        '404': { $ref: '#/components/responses/NotFound' }
  /claims/{claimId}/history:
    parameters: [{ $ref: '#/components/parameters/ClaimId' }]
    get:
      tags: [History]
      operationId: listClaimHistory
      summary: Append-only history, newest first
      description: Mock `listHistory`. Events are never edited or deleted.
      x-slice-status: implemented
      parameters:
        - { name: limit, in: query, schema: { type: integer, minimum: 1, maximum: 200, default: 50 } }
        - $ref: '#/components/parameters/Cursor'
      responses:
        '200':
          description: A page of events.
          content:
            application/json:
              schema:
                type: object
                required: [items, nextCursor]
                properties:
                  items: { type: array, items: { $ref: '#/components/schemas/HistoryEvent' } }
                  nextCursor: { type: [string, 'null'] }
        '400': { $ref: '#/components/responses/BadRequest' }
        '404': { $ref: '#/components/responses/NotFound' }
...
  /claims/{claimId}/workflow-runs:
    parameters: [{ $ref: '#/components/parameters/ClaimId' }]
    get:
      tags: [Workflow runs]
      operationId: listClaimWorkflowRuns
      summary: What each workflow run did
      description: |
        Mock `LifeFlow.runs`. Written by the workflows' own activities, so the record and the state change it describes
        commit together. `swimlane` rows and `serviceLevels` are projections and are not part of this resource.
      x-slice-status: implemented
      responses:
        '200':
          description: Runs, oldest first.
          content:
            application/json:
              schema:
                type: object
                required: [items, nextCursor]
                properties:
                  items: { type: array, items: { $ref: '#/components/schemas/WorkflowRun' } }
                  nextCursor: { type: [string, 'null'] }
        '404': { $ref: '#/components/responses/NotFound' }
...
  /deadlines:check-overdue:
    post:
      tags: [Deadlines]
      operationId: checkOverdueDeadlines
      summary: The nightly overdue check (batch)
      description: |
        Lists open deadline rows past due whose kind has NO workflow (nothing will ever fire them: `acknowledge_by`, `forms_by`, `first_contact_by`,
        `status_letter`, `decision_due`, `review_target`, `payment_due`, `document_review_by`, `bank_details_by`) on claims that are not closed, and, ONCE per
        row, writes a work item for the claim's owner and a history event. It does not close or change the row: the row closes when the thing it waits for
        happens (the welcome call completing `first_contact_by`, the decision, the payment run, the payee's new account). The daily job
        (`Claims:OverdueCheck`) calls the same service; running it again raises nothing new.
      x-slice-status: implemented
      parameters:
        - $ref: '#/components/parameters/Actor'
      responses:
        '200':
          description: What was found.
          content:
            application/json:
              schema:
                type: object
                required: [checkedAt, raised, alreadyRaised, items]
                properties:
                  checkedAt: { type: string, format: date-time }
                  raised: { type: integer, description: Work items written by this call. }
                  alreadyRaised: { type: integer, description: Overdue rows that already had their work item. }
                  items:
                    type: array
                    items:
                      type: object
                      required: [deadlineId, claimId, kind, what, dueAt, raised]
                      properties:
                        deadlineId: { type: string, format: uuid }
                        claimId: { type: string, format: uuid }
                        kind: { $ref: '#/components/schemas/DeadlineKind' }
                        what: { type: string }
                        dueAt: { type: string, format: date-time }
                        raised: { type: boolean, description: True when this call wrote the work item. }
        '400': { $ref: '#/components/responses/BadRequest' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: 'The intended scheme (org identity provider). NOT enforced yet: the service is unauthenticated and trusts X-Actor. See the README, not built yet.'
  headers:
    ETag:
      description: The resource's version, quoted. Send it back as `If-Match`.
      schema: { type: string, examples: ['"3"'] }
  parameters:
    ClaimId: { name: claimId, in: path, required: true, schema: { type: string, format: uuid } }
...
    DeadlineId: { name: deadlineId, in: path, required: true, schema: { type: string, format: uuid } }
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      description: A client-generated unique key (for example a UUID), 8 to 128 characters.
      schema: { type: string, minLength: 8, maxLength: 128 }
    IfMatch:
      name: If-Match
      in: header
      required: true
      description: The ETag from the last GET of this resource.
      schema: { type: string, examples: ['"3"'] }
...
    Actor:
      name: X-Actor
      in: header
      required: false
      description: Development placeholder for the acting user until authentication is built.
      schema: { type: string }
    Limit:
      name: limit
      in: query
      schema: { type: integer, minimum: 1, maximum: 100, default: 25 }
    Cursor:
      name: cursor
      in: query
      description: Opaque; the `nextCursor` of the previous page.
      schema: { type: string }
  responses:
    BadRequest:
      description: The request could not be read (malformed JSON, bad header or cursor).
      content: { application/problem+json: { schema: { $ref: '#/components/schemas/Problem' } } }
...
    NotFound:
      description: No such resource.
      content: { application/problem+json: { schema: { $ref: '#/components/schemas/Problem' } } }
    Conflict:
      description: The request conflicts with the resource's state (`invalid_state`, `idempotency_key_reuse`, `deadline_not_open`, `claim_not_ready_to_decide`, `not_awaiting_approval`, `item_in_run`, ...).
      content: { application/problem+json: { schema: { $ref: '#/components/schemas/Problem' } } }
    PreconditionFailed:
      description: '`If-Match` is stale (`version_conflict`). Fetch the resource again.'
      content: { application/problem+json: { schema: { $ref: '#/components/schemas/Problem' } } }
    PreconditionRequired:
      description: '`If-Match` is required (`precondition_required`).'
      content: { application/problem+json: { schema: { $ref: '#/components/schemas/Problem' } } }
    Unprocessable:
      description: Well formed but not valid (`validation_failed` with field errors, or a business-rule code such as `payee_shares_invalid`).
      content: { application/problem+json: { schema: { $ref: '#/components/schemas/Problem' } } }
  schemas:
...
    Money:
      type: object
      required: [amount, currency]
      properties:
        amount: { type: string, pattern: '^[0-9]{1,13}(\.[0-9]{1,2})?$', examples: ['200000.00'] }
        currency: { type: string, pattern: '^[A-Z]{3}$', examples: [USD] }
    Family: { type: string, enum: [life, disability, annuity] }
    ClaimStatus:
      type: string
      enum: [received, gathering_evidence, in_review, awaiting_approval, approved, paying, closed, reopened]
    DeadlineKind:
      type: string
      description: |
        Only `requirement_follow_up` has a workflow (the dispatcher fires it). Every other kind is a row that something else closes, or a target the
        nightly overdue check raises to a person when it is late: `document_review_by` (1 business day for the examiner to review a document the
        rules could not accept; closed by the review) and `bank_details_by` (a payee's new bank details after a returned payment, 10 days; closed by
        `update-payment-method`) are display/target rows with no automation of their own. `payment_due` also carries the "reissue a returned
        payment" target (`sla: reissue`), closed by the payment run.
      enum: [acknowledge_by, forms_by, first_contact_by, status_letter, requirement_follow_up, decision_due, review_target, payment_due, document_review_by, bank_details_by]
    DeadlineState:
      type: string
      description: open, then dispatched when the dispatcher starts its workflow, then done or skipped.
      enum: [open, dispatched, done, skipped]
...
    PartyRef:
      type: object
      required: [id, name]
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
    LifeIntakeRequest:
      type: object
      required: [caller, insured, death, policies, designation]
      properties:
        noticeReceivedAt: { type: string, format: date-time, description: When the call started. Defaults to the server clock. }
        caller:
          type: object
          required: [name, contactBy, verifiedDateOfBirth, verifiedPolicyNumber]
          properties:
            name: { type: string, maxLength: 200 }
            relationship: { type: string }
            phone: { type: string }
            email: { type: string }
            contactBy:
              type: array
              uniqueItems: true
              items: { type: string, enum: [email, text, phone, mail] }
            verifiedDateOfBirth: { type: boolean, description: Both verification flags must be true (422 `identity_not_verified`). }
            verifiedPolicyNumber: { type: boolean }
            agentConsent: { type: boolean, default: false }
            beneficiaryRef: { type: string, description: 'The ref of the designated beneficiary the caller is, if any.' }
        insured:
          type: object
          required: [name, dateOfBirth]
          properties:
            name: { type: string }
            dateOfBirth: { type: string, format: date }
            ssnLast4: { type: string, pattern: '^[0-9]{4}$' }
        death:
          type: object
          required: [dateOfDeath, manner]
          properties:
            dateOfDeath: { type: string, format: date }
            placeOfDeath: { type: string }
            manner: { type: string, enum: [natural, accident, pending] }
            outsideUs: { type: boolean, default: false }
            funeralHome: { type: string }
        policies:
          type: array
          minItems: 1
          description: The policies being claimed, as found in policy administration. Policy administration is re-checked by the intake workflow.
          items:
            type: object
            required: [policyNumber, productCode, productName, issueDate, paidToDate, inForce, faceAmount]
            properties:
              policyNumber: { type: string }
              productCode: { type: string }
              productName: { type: string }
              issueDate: { type: string, format: date }
              paidToDate: { type: string, format: date }
              inForce: { type: boolean }
              lapsedOn: { type: [string, 'null'], format: date, description: Present exactly when `inForce` is false. }
              faceAmount: { $ref: '#/components/schemas/Money' }
              riders:
                type: array
                items:
                  type: object
                  required: [key, name, amount]
                  properties:
                    key: { type: string, examples: [adb] }
                    name: { type: string }
                    amount: { $ref: '#/components/schemas/Money' }
        designation:
          type: object
          required: [date, beneficiaries]
          properties:
            date: { type: string, format: date }
            source: { type: string }
            beneficiaries:
              type: array
              minItems: 1
              items:
                type: object
                required: [ref, name, kind, sharePercent]
                properties:
                  ref: { type: string, description: 'Client-chosen, unique within the request.' }
                  name: { type: string }
                  relationship: { type: string }
                  kind: { type: string, enum: [primary, contingent] }
                  sharePercent: { type: number, description: 'Percent. The living payees'' shares must add up to 100 (422 `payee_shares_invalid`).' }
                  dateOfBirth: { type: string, format: date }
                  diedOn: { type: string, format: date }
                  diedSource: { type: string }
                  contact:
                    type: object
                    properties:
                      phone: { type: string }
                      email: { type: string }
                      address: { type: string }
                      packet: { type: string, enum: [portal, mail] }
        agent:
          type: object
          required: [name]
          properties:
            name: { type: string }
            agency: { type: string }
        otherClaimantsPossible: { type: boolean, default: false }
...
    Claim:
      type: object
      required: [id, claimNumber, family, productCode, status, routeReasons, team, noticedAt, insured, benefitLines, parties, version, createdAt, updatedAt]
      properties:
        id: { type: string, format: uuid }
        claimNumber: { type: string, pattern: '^[LDA]-[0-9]{2}-[0-9]{6}$', examples: [L-26-043310] }
        family: { $ref: '#/components/schemas/Family' }
        productCode: { type: string }
        status: { $ref: '#/components/schemas/ClaimStatus' }
        track: { type: [string, 'null'], enum: [fast_track_life, standard_life, null], description: Null until the intake workflow has routed the claim. }
        routeRule: { type: [string, 'null'], examples: [LF-01] }
        routeReasons: { type: array, items: { type: string } }
        ownerId: { type: [string, 'null'], format: uuid }
        team: { type: string }
        noticedAt: { type: string, format: date-time }
        proofCompleteAt: { type: [string, 'null'], format: date-time }
        closedAt: { type: [string, 'null'], format: date-time }
        insured: { $ref: '#/components/schemas/PartyRef' }
        details:
          description: Family-specific facts. Only `life` exists in this slice.
          oneOf:
            - $ref: '#/components/schemas/LifeDetails'
          discriminator:
            propertyName: kind
        benefitLines: { type: array, items: { $ref: '#/components/schemas/BenefitLine' } }
        parties: { type: array, items: { $ref: '#/components/schemas/ClaimParty' } }
        version: { type: integer, format: int64 }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    ClaimSummary:
      type: object
      required: [id, claimNumber, family, productCode, status, insuredName, noticedAt, version]
      properties:
        id: { type: string, format: uuid }
        claimNumber: { type: string }
        family: { $ref: '#/components/schemas/Family' }
        productCode: { type: string }
        status: { $ref: '#/components/schemas/ClaimStatus' }
        track: { type: [string, 'null'] }
        ownerId: { type: [string, 'null'], format: uuid }
        insuredName: { type: string }
        noticedAt: { type: string, format: date-time }
        version: { type: integer, format: int64 }
...
    RequirementList:
      type: object
      required: [items, nextCursor]
      properties:
        items: { type: array, items: { $ref: '#/components/schemas/Requirement' } }
        nextCursor: { type: [string, 'null'] }
    Deadline:
      type: object
      required: [id, claimId, kind, what, dueAt, originalDueAt, state, attempt, fired, version]
      properties:
        id: { type: string, format: uuid }
        claimId: { type: string, format: uuid }
        kind: { $ref: '#/components/schemas/DeadlineKind' }
        requirementId: { type: [string, 'null'], format: uuid }
        documentId: { type: [string, 'null'], format: uuid, description: 'The document a `document_review_by` row is about.' }
        partyId: { type: [string, 'null'], format: uuid, description: 'The payee a `bank_details_by` or reissue row is about.' }
        what: { type: string }
        sla: { type: [string, 'null'], description: 'The service level this row enforces: ack, forms, contact, status, decide, review, pay, docReview, reissue.' }
        dueAt: { type: string, format: date-time }
        originalDueAt: { type: string, format: date-time }
        extensionReason: { type: [string, 'null'] }
        state: { $ref: '#/components/schemas/DeadlineState' }
        attempt: { type: integer, description: Dispatch attempts so far. }
        dispatchedAt: { type: [string, 'null'], format: date-time }
        fired: { type: boolean, description: True when it came due and a deadline workflow ran for it. }
        lastOutcome: { type: [string, 'null'], enum: [started, already_started, start_failed, completed, skipped, closed_early, null] }
        lastError: { type: [string, 'null'] }
        workflowId: { type: [string, 'null'], examples: ['deadline-3f2b...'] }
        closedAt: { type: [string, 'null'], format: date-time }
        closedBy: { type: [string, 'null'], enum: [workflow, intake, requirement, decision, payment, user, null] }
        result: { type: [string, 'null'] }
        version: { type: integer, format: int64 }
    DeadlineList:
      type: object
      required: [items, nextCursor]
      properties:
        items: { type: array, items: { $ref: '#/components/schemas/Deadline' } }
        nextCursor: { type: [string, 'null'] }
...
    Letter:
      type: object
      required: [id, claimId, templateCode, channel, recipientLabel, subject, status, createdAt]
      properties:
        id: { type: string, format: uuid }
        claimId: { type: string, format: uuid }
        benefitLineId: { type: [string, 'null'], format: uuid }
        templateCode: { type: string, examples: [ACK-LIFE-01] }
        channel: { type: string, enum: [letter, email, sms, portal, fax] }
        recipientPartyId: { type: [string, 'null'], format: uuid }
        recipientLabel: { type: string }
        subject: { type: string }
        summary: { type: [string, 'null'] }
        status:
          type: string
          enum: [draft, queued, sent, failed, skipped]
          description: |
            `skipped` is new: the letter was queued but the situation changed before it went (for example a re-run of the document workflow accepted the
            W-9 that run 1 had asked to be corrected), so it will never be sent; `statusReason` says why. A workflow that has finished leaves no letter `queued`.
        statusReason: { type: [string, 'null'], description: 'Why the letter is `skipped`; null otherwise. Additive.' }
        sentAt: { type: [string, 'null'], format: date-time }
        workflowId: { type: [string, 'null'] }
        createdAt: { type: string, format: date-time }
...